+++ src/libs/ajax/Runtimes.js
@@ -1,7 +1,7 @@
 import _ from 'lodash/fp'
 import * as qs from 'qs'
 import { version } from 'src/data/machines'
-import { appIdentifier, authOpts, fetchLeo, fetchOk, jsonBody } from 'src/libs/ajax'
+import { appIdentifier, authOpts, fetchLeo, fetchOk, jsonBody } from 'src/libs/ajax/ajax-common'
 import { getConfig } from 'src/libs/config'
