+++ src/libs/auth.js
@@ -5,7 +5,8 @@ import { Fragment } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
 import { FrameworkServiceLink, ShibbolethLink, UnlinkFenceAccount } from 'src/components/common'
 import { cookiesAcceptedKey } from 'src/components/CookieWarning'
-import { Ajax, fetchOk } from 'src/libs/ajax'
+import { Ajax } from 'src/libs/ajax'
+import { fetchOk } from 'src/libs/ajax/ajax-common'
 import { getEnabledBrand } from 'src/libs/brand-utils'
 import { getConfig } from 'src/libs/config'
 import { withErrorIgnoring, withErrorReporting } from 'src/libs/error'
